controllers: extract common resource getters to common package
